alles geupdate

+++ 05_meer_functions/calculate/functions.py
@@ -1,16 +1,16 @@
-def addition(n1 : float, n2 : float):
+def addition(n1 : int, n2 : int):
     answer = n1 + n2
     return answer
 
-def subtraction(n1 : float, n2 : float):
+def subtraction(n1 : int, n2 : int):
     answer = n1 - n2
     return answer
 
-def multiplication(n1 : float, n2 : float):
+def multiplication(n1 : int, n2 : int):
     answer = n1 * n2
     return answer
 
-def division(n1 : float, n2 : float):
+def division(n1 : int, n2 : int):
     answer = n1 / n2
     return answer
